Fix: delete comment

+++ src/main/java/com/habibInc/issueTracker/comment/CommentService.java
@@ -38,11 +38,14 @@ public class CommentService {
 
     public void deleteComment(Long issueId, Long commentId, User authenticatedUser) {
         // find the comment by its issue's id (ensure that both the issue and comment exist)
-        Comment comment = getCommentById(issueId);
+        Comment comment = getCommentById(commentId);
 
         // in case the owner is not the authenticated user, throw a forbidden operation error
         if(!comment.getOwner().equals(authenticatedUser))
             throw new ForbiddenOperationException("Forbidden");
+
+        if(comment.getIssue().getId() != issueId)
+            throw new ResourceNotFoundException("Issue not found");
 
         // delete the comment if it exists, otherwise a ResourceNotFoundException is thrown
         commentRepository.deleteById(commentId);
